feat(149): separate responsibilities and decouple methods
