+++ app/src/androidTest/java/org/cescfe/numpairs/feature/fourpairs/FourPairsActionDiscoveryTest.kt
@@ -6,12 +6,7 @@ import androidx.compose.ui.test.junit4.v2.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.update
-import org.cescfe.numpairs.data.preferences.TopAppBarActionDiscoveryRepository
-import org.cescfe.numpairs.data.preferences.TopAppBarActionDiscoveryState
+import org.cescfe.numpairs.data.preferences.FakeTopAppBarActionDiscoveryRepository
 import org.cescfe.numpairs.data.puzzle.seed.initialPuzzle
 import org.cescfe.numpairs.feature.game.ui.GameScreenTestTags
 import org.cescfe.numpairs.ui.theme.NumPairsTheme
@@ -73,28 +68,6 @@ class FourPairsActionDiscoveryTest {
         composeTestRule.runOnIdle {
             assertTrue(actionDiscoveryRepository.state.value.hasSeenHelpAction)
             assertTrue(actionDiscoveryRepository.state.value.hasSeenHintAction)
-        }
-    }
-
-    private class FakeTopAppBarActionDiscoveryRepository(
-        initialState: TopAppBarActionDiscoveryState = TopAppBarActionDiscoveryState()
-    ) : TopAppBarActionDiscoveryRepository {
-        private val mutableState = MutableStateFlow(initialState)
-
-        val state: StateFlow<TopAppBarActionDiscoveryState> = mutableState.asStateFlow()
-
-        override val discoveryState = state
-
-        override suspend fun markHelpActionSeen() {
-            mutableState.update { state ->
-                state.copy(hasSeenHelpAction = true)
-            }
-        }
-
-        override suspend fun markHintActionSeen() {
-            mutableState.update { state ->
-                state.copy(hasSeenHintAction = true)
-            }
         }
     }
 }
