+++ app/src/main/java/org/cescfe/numpairs/MainActivity.kt
@@ -5,6 +5,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import org.cescfe.numpairs.data.preferences.createTopAppBarActionDiscoveryRepository
 import org.cescfe.numpairs.ui.navigation.AppNavigation
 import org.cescfe.numpairs.ui.theme.NumPairsTheme
 
@@ -13,9 +14,13 @@ class MainActivity : ComponentActivity() {
         installSplashScreen()
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
+        val topAppBarActionDiscoveryRepository = createTopAppBarActionDiscoveryRepository(applicationContext)
+
         setContent {
             NumPairsTheme {
-                AppNavigation()
+                AppNavigation(
+                    topAppBarActionDiscoveryRepository = topAppBarActionDiscoveryRepository
+                )
             }
         }
     }
