+++ app/src/main/java/org/cescfe/numpairs/data/preferences/TopAppBarActionDiscoveryRepositoryFactory.kt
@@ -1,0 +1,6 @@
+package org.cescfe.numpairs.data.preferences
+
+import android.content.Context
+
+fun createTopAppBarActionDiscoveryRepository(context: Context): TopAppBarActionDiscoveryRepository =
+    DataStoreTopAppBarActionDiscoveryRepository(context.applicationContext.userPreferencesDataStore)
