+++ app/src/main/java/org/cescfe/numpairs/data/preferences/UserPreferencesDataStore.kt
@@ -1,0 +1,12 @@
+package org.cescfe.numpairs.data.preferences
+
+import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.preferencesDataStore
+
+private const val USER_PREFERENCES_DATA_STORE_NAME = "user_preferences"
+
+internal val Context.userPreferencesDataStore: DataStore<Preferences> by preferencesDataStore(
+    name = USER_PREFERENCES_DATA_STORE_NAME
+)
